Removed unnecessary float -> int conversions in rendering code

+++ TwistedLogik.Ultraviolet/Graphics/Graphics2D/StretchableImage3.cs
@@ -225,7 +225,7 @@ namespace TwistedLogik.Ultraviolet.Graphics.Graphics2D
         }
 
         /// <inheritdoc/>
-        internal override void Draw<VertexType, SpriteData>(SpriteBatchBase<VertexType, SpriteData> spriteBatch, Vector2 position, Int32 width, Int32 height, Color color, Single rotation, Vector2 origin, SpriteEffects effects, Single layerDepth, SpriteData data)
+        internal override void Draw<VertexType, SpriteData>(SpriteBatchBase<VertexType, SpriteData> spriteBatch, Vector2 position, Single width, Single height, Color color, Single rotation, Vector2 origin, SpriteEffects effects, Single layerDepth, SpriteData data)
         {
             effects |= SpriteEffects.OriginRelativeToDestination;
 
@@ -242,7 +242,7 @@ namespace TwistedLogik.Ultraviolet.Graphics.Graphics2D
         /// <summary>
         /// Draws vertical images.
         /// </summary>
-        private void DrawVertical<VertexType, SpriteData>(SpriteBatchBase<VertexType, SpriteData> spriteBatch, Vector2 position, Int32 width, Int32 height, Color color, Single rotation, Vector2 origin, SpriteEffects effects, Single layerDepth, SpriteData data)
+        private void DrawVertical<VertexType, SpriteData>(SpriteBatchBase<VertexType, SpriteData> spriteBatch, Vector2 position, Single width, Single height, Color color, Single rotation, Vector2 origin, SpriteEffects effects, Single layerDepth, SpriteData data)
             where VertexType : struct, IVertexType
             where SpriteData : struct
         {
@@ -253,7 +253,7 @@ namespace TwistedLogik.Ultraviolet.Graphics.Graphics2D
 
             if (height < MinimumRecommendedSize.Height)
             {
-                var scale = height / (float)MinimumRecommendedSize.Height;
+                var scale = height / MinimumRecommendedSize.Height;
                 dstTop    = (Int32)Math.Floor(dstTop * scale);
                 dstBottom = (Int32)Math.Ceiling(dstBottom * scale);
             }
@@ -307,7 +307,7 @@ namespace TwistedLogik.Ultraviolet.Graphics.Graphics2D
         /// <summary>
         /// Draws horizontal images.
         /// </summary>
-        private void DrawHorizontal<VertexType, SpriteData>(SpriteBatchBase<VertexType, SpriteData> spriteBatch, Vector2 position, Int32 width, Int32 height, Color color, Single rotation, Vector2 origin, SpriteEffects effects, Single layerDepth, SpriteData data)
+        private void DrawHorizontal<VertexType, SpriteData>(SpriteBatchBase<VertexType, SpriteData> spriteBatch, Vector2 position, Single width, Single height, Color color, Single rotation, Vector2 origin, SpriteEffects effects, Single layerDepth, SpriteData data)
             where VertexType : struct, IVertexType
             where SpriteData : struct
         {
@@ -318,7 +318,7 @@ namespace TwistedLogik.Ultraviolet.Graphics.Graphics2D
 
             if (width < MinimumRecommendedSize.Width)
             {
-                var scale = width / (float)MinimumRecommendedSize.Width;
+                var scale = width / MinimumRecommendedSize.Width;
                 dstLeft   = (Int32)Math.Floor(dstLeft * scale);
                 dstRight  = (Int32)Math.Ceiling(dstRight * scale);
             }
